% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-aoc.R
\name{labeller_aoc}
\alias{labeller_aoc}
\title{A labeller function to supply to facet_grid() or facet_wrap() for the argument labeller.}
\usage{
labeller_aoc(...)
}
\arguments{
\item{...}{optional parameters to the labeller function}
}
\description{
A labeller function to supply to facet_grid() or facet_wrap() for the argument labeller.
}
