% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-aoc.R
\name{labs_aoc}
\alias{labs_aoc}
\title{Set a Alexandria Ocasio Cortez inspired look to a ggplot2 labels}
\usage{
labs_aoc(title, ...)
}
\arguments{
\item{title}{that it will be transformed uppercase}

\item{...}{additional parameters to the labs function}
}
\description{
Set a Alexandria Ocasio Cortez inspired look to a ggplot2 labels
}
