% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-aoc.R
\name{scale_color_aoc}
\alias{scale_color_aoc}
\alias{scale_colour_aoc}
\title{Set a Alexandria Ocasio Cortez inspired look to a ggplot2 color scales}
\usage{
scale_color_aoc()

scale_colour_aoc()
}
\description{
Set a Alexandria Ocasio Cortez inspired look to a ggplot2 color scales
}
